+++ core/source/driver/PrimitiveRegion.cpp
@@ -11,14 +11,14 @@ import tensor;
 export struct PrimitiveRegion : ComputeTool
 {
 	PrimitiveRegion(const std::string& name,MetricSpace *ms,Task *tsk) : ComputeTool(name,ms,tsk) {}
-	virtual bool Inside(const tw::vec3& pos) const = 0;
+	virtual bool Inside(const tw::vec4& pos) const = 0;
     virtual std::array<tw::Float,6> Bounds() const = 0;
 };
 
 export struct EntireRegion : PrimitiveRegion
 {
 	EntireRegion(const std::string& name,MetricSpace *ms,Task *tsk) : PrimitiveRegion(name,ms,tsk) {}
-	virtual bool Inside(const tw::vec3& pos) const {
+	virtual bool Inside(const tw::vec4& pos) const {
 		return true;
 	}
     virtual std::array<tw::Float,6> Bounds() const {
@@ -32,8 +32,8 @@ export struct RectRegion : PrimitiveRegion
 	RectRegion(const std::string& name,MetricSpace *ms,Task *tsk) : PrimitiveRegion(name,ms,tsk) {
 		directives.Add("size",new tw::input::Vec3(&size),true);
 	}
-	virtual bool Inside(const tw::vec3& p) const {
-		return std::fabs(2*p.x)<size.x && std::fabs(2*p.y)<size.y && std::fabs(2*p.z)<size.z;
+	virtual bool Inside(const tw::vec4& p) const {
+		return std::fabs(2*p[1])<size.x && std::fabs(2*p[2])<size.y && std::fabs(2*p[3])<size.z;
 	}
     virtual std::array<tw::Float,6> Bounds() const {
 		return std::array<tw::Float,6> {-size.x/2,size.x/2,-size.y/2,size.y/2,-size.z/2,size.z/2};
@@ -43,8 +43,8 @@ export struct RectRegion : PrimitiveRegion
 export struct PrismRegion : RectRegion
 {
 	PrismRegion(const std::string& name,MetricSpace *ms,Task *tsk) : RectRegion(name,ms,tsk) {}
-	virtual bool Inside(const tw::vec3& p) const {
-		return std::fabs(2*p.x)<size.x && std::fabs(2*p.y)<size.y && std::fabs(2*p.z)<size.z*0.5*(size.x-2*p.x)/size.x;
+	virtual bool Inside(const tw::vec4& p) const {
+		return std::fabs(2*p[1])<size.x && std::fabs(2*p[2])<size.y && std::fabs(2*p[3])<size.z*0.5*(size.x-2*p[1])/size.x;
 	}
 };
 
@@ -55,7 +55,7 @@ export struct CircRegion : PrimitiveRegion
 		radius = 1.0;
 		directives.Add("radius",new tw::input::Float(&radius),true);
 	}
-	virtual bool Inside(const tw::vec3& p) const {
+	virtual bool Inside(const tw::vec4& p) const {
 		return Norm(p)<sqr(radius);
 	}
     virtual std::array<tw::Float,6> Bounds() const {
@@ -70,8 +70,8 @@ export struct CylinderRegion : PrimitiveRegion
 		directives.Add("radius",new tw::input::Float(&radius),true);
 		directives.Add("length",new tw::input::Float(&length),true);
 	}
-	virtual bool Inside(const tw::vec3& p) const {
-		return sqr(p.x) + sqr(p.y) < sqr(radius) && std::fabs(2*p.z) < length;
+	virtual bool Inside(const tw::vec4& p) const {
+		return sqr(p[1]) + sqr(p[2]) < sqr(radius) && std::fabs(2*p[3]) < length;
 	}
     virtual std::array<tw::Float,6> Bounds() const {
 		return std::array<tw::Float,6> {-radius,radius,-radius,radius,-length/2,length/2};
@@ -86,10 +86,10 @@ export struct CylindricalShellRegion : PrimitiveRegion
 		directives.Add("outer radius",new tw::input::Float(&outerRadius),true);
 		directives.Add("length",new tw::input::Float(&length),true);
 	}
-	virtual bool Inside(const tw::vec3& p) const {
+	virtual bool Inside(const tw::vec4& p) const {
 		tw::Float rho;
-		rho = std::sqrt(p.x*p.x + p.y*p.y);
-		return rho < outerRadius && rho > innerRadius && std::fabs(2*p.z) < length;
+		rho = std::sqrt(p[1]*p[1] + p[2]*p[2]);
+		return rho < outerRadius && rho > innerRadius && std::fabs(2*p[3]) < length;
 	}
     virtual std::array<tw::Float,6> Bounds() const {
 		return std::array<tw::Float,6> {-outerRadius,outerRadius,-outerRadius,outerRadius,-length/2,length/2};
@@ -100,11 +100,11 @@ export struct RoundedCylinderRegion : CylinderRegion
 {
 	RoundedCylinderRegion(const std::string& name,MetricSpace *ms,Task *tsk) : CylinderRegion(name,ms,tsk) {}
 	// TODO: the inherited Bounds function does not enclose everything
-	virtual bool Inside(const tw::vec3& p) const {
+	virtual bool Inside(const tw::vec4& p) const {
 		bool ans;
-		ans = sqr(p.x) + sqr(p.y) < sqr(radius) && std::fabs(2*p.z) < length;
-		ans = ans || Norm(p - tw::vec3(0,0,length/2)) < sqr(radius);
-		ans = ans || Norm(p + tw::vec3(0,0,length/2)) < sqr(radius);
+		ans = sqr(p[1]) + sqr(p[2]) < sqr(radius) && std::fabs(2*p[3]) < length;
+		ans = ans || Norm(p - tw::vec4(0,0,0,length/2)) < sqr(radius);
+		ans = ans || Norm(p + tw::vec4(0,0,0,length/2)) < sqr(radius);
 		return ans;
 	}
 };
@@ -112,15 +112,15 @@ export struct RoundedCylinderRegion : CylinderRegion
 export struct EllipsoidRegion : RectRegion
 {
 	EllipsoidRegion(const std::string& name,MetricSpace *ms,Task *tsk) : RectRegion(name,ms,tsk) {}
-	virtual bool Inside(const tw::vec3& p) const {
-		return sqr(2*p.x/size.x) + sqr(2*p.y/size.y) + sqr(2*p.z/size.z) < 1.0;
+	virtual bool Inside(const tw::vec4& p) const {
+		return sqr(2*p[1]/size.x) + sqr(2*p[2]/size.y) + sqr(2*p[3]/size.z) < 1.0;
 	}
 };
 
 // export struct TrueSphere : CircRegion
 // {
 // 	TrueSphere(const std::string& name,MetricSpace *ms,Task *tsk) : CircRegion(name,ms,tsk) {}
-// 	virtual bool Inside(const tw::vec3& pos,int depth) const {
+// 	virtual bool Inside(const tw::vec4& pos,int depth) const {
 // 		tw::vec3 c_cart = translation;
 // 		tw::vec3 p_cart = pos;
 // 		space->CurvilinearToCartesian(&c_cart);
@@ -140,13 +140,14 @@ export struct BoxArrayRegion : PrimitiveRegion
 		directives.Add("size",new tw::input::Vec3(&size),true);
 		directives.Add("spacing",new tw::input::Vec3(&spacing),true);
 	}
-	virtual bool Inside(const tw::vec3& pos) const {
+	virtual bool Inside(const tw::vec4& pos) const {
 		// this is an infinite array, but of course can be intersected with something else to produce a finite one
-		auto p = pos + 0.5*size;
-		tw::Int i = MyFloor(p.x/spacing.x);
-		tw::Int j = MyFloor(p.y/spacing.y);
-		tw::Int k = MyFloor(p.z/spacing.z);
-		return p.x - tw::Float(i)*spacing.x < size.x && p.y - tw::Float(j)*spacing.y < size.y && p.z - tw::Float(k)*spacing.z < size.z;
+		auto p = pos;
+		p.Add3(0.5*size);
+		tw::Int i = MyFloor(p[1]/spacing.x);
+		tw::Int j = MyFloor(p[2]/spacing.y);
+		tw::Int k = MyFloor(p[3]/spacing.z);
+		return p[1] - tw::Float(i)*spacing.x < size.x && p[2] - tw::Float(j)*spacing.y < size.y && p[3] - tw::Float(k)*spacing.z < size.z;
 	}
     virtual std::array<tw::Float,6> Bounds() const {
 		return std::array<tw::Float,6> {-size.x/2,size.x/2,-size.y/2,size.y/2,-size.z/2,size.z/2};
@@ -163,11 +164,11 @@ export struct TorusRegion : PrimitiveRegion
 		directives.Add("major radius",new tw::input::Float(&majorRadius));
 		directives.Add("length",new tw::input::Float(&length),true);
 	}
-	virtual bool Inside(const tw::vec3& p) const {
+	virtual bool Inside(const tw::vec4& p) const {
 		tw::Float rho;
-		rho = std::sqrt(p.x*p.x + p.y*p.y);
+		rho = std::sqrt(p[1]*p[1] + p[2]*p[2]);
 		return rho > majorRadius-minorRadius && rho < majorRadius+minorRadius &&
-			sqr(p.z) < sqr(minorRadius)-sqr(rho-majorRadius);
+			sqr(p[3]) < sqr(minorRadius)-sqr(rho-majorRadius);
 	}
     virtual std::array<tw::Float,6> Bounds() const {
 		auto r = tw::vec3(majorRadius+minorRadius,majorRadius+minorRadius,minorRadius);
@@ -186,11 +187,11 @@ export struct ConeRegion : PrimitiveRegion
 		directives.Add("base radius",new tw::input::Float(&baseRadius));
 		directives.Add("length",new tw::input::Float(&length),true);
 	}
-	virtual bool Inside(const tw::vec3& p) const {
+	virtual bool Inside(const tw::vec4& p) const {
 		tw::Float rho,rOfz;
-		rho = std::sqrt(p.x*p.x + p.y*p.y);
-		rOfz = tipRadius - (p.z - length/2)*(baseRadius-tipRadius)/length;
-		return rho < rOfz && sqr(p.z) < sqr(length/2);
+		rho = std::sqrt(p[1]*p[1] + p[2]*p[2]);
+		rOfz = tipRadius - (p[3] - length/2)*(baseRadius-tipRadius)/length;
+		return rho < rOfz && sqr(p[3]) < sqr(length/2);
 	}
     virtual std::array<tw::Float,6> Bounds() const {
 		auto r = tw::vec3(baseRadius,baseRadius,length/2);
@@ -206,7 +207,7 @@ export struct TangentOgiveRegion : PrimitiveRegion
 		directives.Add("body radius",new tw::input::Float(&bodyRadius),true);
 		directives.Add("length",new tw::input::Float(&length),true);
 	}
-	virtual bool Inside(const tw::vec3& p) const {
+	virtual bool Inside(const tw::vec4& p) const {
 		tw::Float rho,rOfz,ogiveRadius,x0,xt,yt;
 
 		x0 = length - tipRadius;
@@ -214,8 +215,8 @@ export struct TangentOgiveRegion : PrimitiveRegion
 		yt = tipRadius*(ogiveRadius-bodyRadius)/(ogiveRadius-tipRadius);
 		xt = x0 + std::sqrt(sqr(tipRadius)-yt*yt);
 
-		rho = std::sqrt(p.x*p.x + p.y*p.y);
-		auto z = p.z + length/2;
+		rho = std::sqrt(p[1]*p[1] + p[2]*p[2]);
+		auto z = p[3] + length/2;
 		rOfz = -1.0;
 		if (z>0.0 && z<xt)
 			rOfz = std::sqrt(sqr(ogiveRadius)-z*z) + bodyRadius - ogiveRadius;